+++ testing/test_integration.py
@@ -216,7 +216,7 @@ version = '{version}'
 major = {version_tuple[0]}
 minor = {version_tuple[1]}
 patch = {version_tuple[2]}
-commit_hash = '{scm_version.node}'
+commit_hash = '{scm_version.short_node}'
 num_commit = {scm_version.distance}
 """  # noqa: RUF027
     # Use write_to with template to create version file
@@ -294,7 +294,7 @@ def test_pretend_metadata_with_scm_version(
     # This is a template string, not an f-string - used by setuptools-scm templating
     version_file_content = """
 version = '{version}'
-commit_hash = '{scm_version.node}'
+commit_hash = '{scm_version.short_node}'
 num_commit = {scm_version.distance}
 """  # noqa: RUF027
     version = wd.get_version(